delete unnecessary files
